+++ assets/.cryptotux/scripts/ttyd.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+/home/bobby/.cryptotux/