+++ assets/.cryptotux/scripts/versions.sh
@@ -7,6 +7,7 @@ echo ${vv#"Description:"}
 
 echo -e "\n Languages"
 # python -V # deprecated
+gcc --version | head -n 1
 python3 -V
 go version
 rustc --version 
@@ -16,6 +17,7 @@ echo -e "\n Tooling"
 docker --version
 echo "geth" $(geth version | sed -n 2p)
 bitcoind --version | head -n 1
+# Optionnal installs
+[ -x "$(command -v tendermint)" ] && tendermint version
 
-echo ">> done <<"
-
+echo ">>> done <<<"